Add uv-based CI with SciPy regression checks
