+++ .gitignore
@@ -1,0 +1,23 @@
+# Python artifacts
+__pycache__/
+*.py[cod]
+*$py.class
+
+# Distribution / packaging
+build/
+dist/
+*.egg-info/
+.eggs/
+
+# Virtual environments
+.venv/
+venv/
+.env/
+
+# PyInstaller output
+*.spec
+
+# IDE / editor settings
+.vscode/
+.idea/
+.DS_Store
