+++ pyproject.toml
@@ -1,7 +1,7 @@
 [project]
 name = "how-many"
 version = "0.1.0"
-description = "Count how many things are on a screen."
+description = "Always-on-top overlay for counting repeated image features."
 requires-python = ">=3.10"
 dependencies = [
     "PySide6>=6.5",
@@ -10,14 +10,35 @@ dependencies = [
 ]
 
 [dependency-groups]
-dev = ["pyinstaller>=6.14"]
+build = [
+    "pyinstaller>=6.14",
+]
+test = [
+    "pytest>=8.3",
+    "scipy==1.15.3",
+    "tomli>=2.0.1",
+]
+
+[project.optional-dependencies]
+dev = [
+    "pyinstaller>=6.14",
+    "pytest>=8.3",
+    "scipy==1.15.3",
+    "tomli>=2.0.1",
+]
 
 [project.scripts]
-build-exe = "how_many_build:main"
+how-many = "how_many.app:main"
 
 [tool.uv]
-default-groups = ["dev"]
+default-groups = ["build"]
 package = true
 
 [tool.setuptools]
-py-modules = ["how_many_build"]
+package-dir = {"" = "src"}
+
+[tool.setuptools.packages.find]
+where = ["src"]
+
+[tool.how-many]
+scipy-reference-version = "1.15.3"
