+++ src/how_many/__main__.py
@@ -1,0 +1,9 @@
+"""Module entry point allowing ``python -m how_many``."""
+
+from __future__ import annotations
+
+from .app import main
+
+
+if __name__ == "__main__":  # pragma: no cover
+    main()
