+++ src/how_many/app.py
@@ -1,235 +1,21 @@
-"""
-how_many.py — overlay tool to estimate repeated structures along a line
--------------------------------------------------------------------------------
-Summary
-  • Always‑on‑top overlay draws a line with two draggable dots.
-  • You position it over an image (semiconductor / industrial / biological, etc.).
-  • It analyzes a configurable‑width stripe under the line and estimates the
-    number of repeated items based on periodicity (FFT + autocorrelation).
-  • Presents several candidate counts in a drop‑down; you can also set the count
-    manually with up/down arrows. Selecting a candidate updates the manual box.
-  • Draws tick markers along the line for visual confirmation.
-  • Profile tab shows the 1‑D stripe profile and candidate results.
-  • Hold **Ctrl** while dragging an endpoint to snap the line angle to 0°, 45°, 90°, 135°.
-Defaults
-  • Auto‑analyze is OFF by default (no analysis on move/drop unless enabled).
-  • Tick markers always **include endpoints** (not configurable).
--------------------------------------------------------------------------------
-"""
+"""Qt application entry point for the how_many overlay tool."""
 
-from dataclasses import dataclass, field, asdict
+from __future__ import annotations
+
 from pathlib import Path
-from typing import List, Tuple, Optional, Dict
+from typing import List, Optional, Tuple
 
-import sys
-import json
 import math
+import sys
 
-import numpy as np
 import cv2  # opencv-python
-from peaks_detrend_standalone import find_peaks, detrend
-
+import numpy as np
 from PySide6 import QtCore, QtGui, QtWidgets
 
-
-# ------------------------------ Data Models ----------------------------------
-
-
-@dataclass
-class AnalysisParams:
-    stripe_width_px: int = 20
-    blur_sigma_px: float = 0.0
-    auto_analyze: bool = False  # default OFF
-    hide_during_capture_ms: int = 40
-    suggest_max: int = 5
-    min_length_px: int = 48
-
-
-@dataclass
-class UIState:
-    manual_count: int = 10
-    always_on_top: bool = True
-    line_thickness: int = 3
-    tick_length_px: int = 10
-
-
-@dataclass
-class Suggestion:
-    count: int  # items (endpoints included)
-    confidence: float  # 0..1
-    source: str  # "fft", "acf", "harmonic", etc.
-
-
-@dataclass
-class AppConfig:
-    params: AnalysisParams = field(default_factory=AnalysisParams)
-    ui: UIState = field(default_factory=UIState)
-
-    def to_json(self) -> str:
-        return json.dumps(asdict(self), indent=2)
-
-    @staticmethod
-    def from_json(text: str) -> "AppConfig":
-        data: Dict = json.loads(text)
-        p = data.get("params", {})
-        u = data.get("ui", {})
-        return AppConfig(
-            params=AnalysisParams(
-                stripe_width_px=int(p.get("stripe_width_px", 20)),
-                blur_sigma_px=float(p.get("blur_sigma_px", 0.0)),
-                auto_analyze=bool(p.get("auto_analyze", False)),
-                hide_during_capture_ms=int(p.get("hide_during_capture_ms", 40)),
-                suggest_max=int(p.get("suggest_max", 5)),
-                min_length_px=int(p.get("min_length_px", 48)),
-            ),
-            ui=UIState(
-                manual_count=int(u.get("manual_count", 10)),
-                always_on_top=bool(u.get("always_on_top", True)),
-                line_thickness=int(u.get("line_thickness", 3)),
-                tick_length_px=int(u.get("tick_length_px", 10)),
-            ),
-        )
-
-
-# ----------------------------- Helper Utilities -------------------------------
-
-
-def qpixmap_to_bgr(pix: QtGui.QPixmap) -> np.ndarray:
-    """Convert a QPixmap to an OpenCV BGR numpy array (PySide6-safe)."""
-    fmt = getattr(QtGui.QImage, "Format_RGBA8888", None)
-    if fmt is None:
-        fmt = QtGui.QImage.Format.Format_RGBA8888
-    img: QtGui.QImage = pix.toImage().convertToFormat(fmt)
-    width = img.width()
-    height = img.height()
-    bytes_per_line = img.bytesPerLine()
-    buf = img.constBits()  # memoryview in PySide6
-    arr = np.frombuffer(buf, np.uint8)
-    arr = arr.reshape((height, bytes_per_line))  # include stride
-    arr = arr[:, : width * 4]  # crop padding
-    arr = arr.reshape((height, width, 4))
-    bgr = cv2.cvtColor(arr, cv2.COLOR_RGBA2BGR)
-    return bgr
-
-
-def rotate_point(
-    x: float, y: float, cx: float, cy: float, angle_deg: float
-) -> Tuple[float, float]:
-    theta = math.radians(angle_deg)
-    cos_t = math.cos(theta)
-    sin_t = math.sin(theta)
-    x0 = x - cx
-    y0 = y - cy
-    xr = x0 * cos_t - y0 * sin_t + cx
-    yr = x0 * sin_t + y0 * cos_t + cy
-    return xr, yr
-
-
-def clamp(v: float, lo: float, hi: float) -> float:
-    return max(lo, min(hi, v))
-
-
-# ------------------------------ Core Analysis ---------------------------------
-
-
-def _rank_and_cap(
-    suggestions: Dict[int, Suggestion], profile_len: int, max_candidates: int
-) -> List[Suggestion]:
-    ranked = sorted(
-        suggestions.values(),
-        key=lambda s: (s.confidence, s.source.startswith("fft")),
-        reverse=True,
-    )
-    filtered: List[Suggestion] = [
-        s for s in ranked if 1 <= s.count <= max(1, profile_len)
-    ]  # items can be up to length
-    return filtered[: max(1, int(max_candidates))]
-
-
-def estimate_counts_from_profile(
-    profile: np.ndarray, max_candidates: int = 5
-) -> List[Suggestion]:
-    """
-    Estimate number of ITEMS along the line (including endpoints) using FFT & autocorrelation.
-    Internally, spectral/ACF give an estimate of CYCLES (intervals). We convert via: items = cycles + 1.
-    """
-    N = int(profile.size)
-    if N < 16:
-        return []
-
-    # Detrend and normalize
-    p = detrend(profile.astype(np.float64))
-    m = float(np.max(np.abs(p)))
-    if m > 1e-9:
-        p /= m
-
-    # Windowing to reduce spectral leakage
-    win = np.hanning(N)
-    p_win = p * win
-
-    suggestions: Dict[int, Suggestion] = {}
-
-    def add_items(cycles: int, conf: float, source: str) -> None:
-        # Convert cycles (intervals) -> items (endpoints included)
-        items = int(max(1, cycles + 1))
-        prev = suggestions.get(items)
-        if (prev is None) or (conf > prev.confidence):
-            suggestions[items] = Suggestion(
-                count=items, confidence=float(clamp(conf, 0.0, 1.0)), source=source
-            )
-
-    # ---------------- FFT ----------------
-    fft_mag = np.abs(np.fft.rfft(p_win))
-    fft_mag[0] = 0.0  # ignore DC
-
-    spectral = fft_mag.copy()
-    spectral[:2] = 0.0  # ignore ultra-low frequencies
-
-    spec_max = float(np.max(spectral)) if np.max(spectral) > 0 else 0.0
-    if spec_max > 0:
-        prominence = float(spec_max * 0.15)
-        peaks, _ = find_peaks(spectral, prominence=prominence)
-
-        if peaks.size > 0:
-            for k in peaks:
-                cycles = int(k)  # DFT bin index ~ cycles across the sampled window
-                conf = float(spectral[k] / spec_max)
-                add_items(cycles, conf, "fft-peak")
-
-            # Consider harmonics
-            top_idxs = sorted(
-                peaks.tolist(), key=lambda i: float(spectral[i]), reverse=True
-            )[:3]
-            for i in top_idxs:
-                base = int(i)
-                base_conf = float(spectral[i] / spec_max)
-                add_items(base, base_conf, "fft-top")
-                if 2 * base < len(spectral):
-                    add_items(2 * base, base_conf * 0.6, "harmonic-x2")
-                if base % 2 == 0 and base > 0:
-                    add_items(base // 2, base_conf * 0.6, "harmonic-x0.5")
-
-    # ---------------- ACF ----------------
-    p0 = p - float(np.mean(p))
-    acf_full = np.correlate(p0, p0, mode="full")
-    acf = acf_full[N - 1 :]  # lags 0..N-1
-    if acf[0] != 0:
-        acf = acf / float(acf[0])
-    acf[:2] = 0.0  # ignore lag 0/1
-
-    if N >= 32 and np.max(acf) > 0:
-        acf_peaks, _ = find_peaks(acf, prominence=float(np.max(acf) * 0.1))
-        if acf_peaks.size > 0:
-            acf_max = float(np.max(acf))
-            for lag in acf_peaks[:5]:
-                if lag > 0:
-                    cycles = int(
-                        round((N - 1) / float(lag))
-                    )  # use N-1 (interval length) for cycles
-                    conf = float(acf[lag] / acf_max)
-                    add_items(cycles, conf, "acf-peak")
-
-    return _rank_and_cap(suggestions, N, max_candidates)
+from .analysis import estimate_counts_from_profile
+from .models import AppConfig, Suggestion, UIState
+from .utils import clamp, rotate_point
+from .utils.qt import qpixmap_to_bgr
 
 
 # ------------------------------ Overlay Widget --------------------------------
