+++ src/how_many/utils/__init__.py
@@ -1,0 +1,5 @@
+"""Utility helpers for the how_many application."""
+
+from .geometry import clamp, rotate_point
+
+__all__ = ["clamp", "rotate_point"]
